MVC restructure
Restructure project with MVC ideology
